move constants to pkg/constants
