+++ pkg/controllers/pod_controller.go
@@ -32,7 +32,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
 
-	"sigs.k8s.io/jobset/pkg/controllers/constants"
+	"sigs.k8s.io/jobset/pkg/constants"
 	"sigs.k8s.io/jobset/pkg/util/placement"
 
 	jobset "sigs.k8s.io/jobset/api/jobset/v1alpha2"
